No more pyaudio in klassifier.py
